Refactor: Add ShapeUtils and UserUtils to reduce the ConnectionHub class size
